# backend/src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Bootcamp Sysmap Fullstack 2025 API",
    "version": "1.0.0",
    "description": "Documentação completa da API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor Local"
    }
  ],
  "paths": {
...
    "/activities/types": {
      "get": {
        "summary": "Listar tipos de atividades",
        "tags": ["Atividades"],
        "description": "Endpoint para listar os tipos de atividades disponíveis.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "name": { "type": "string" },
                      "description": { "type": "string" },
                      "image": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/activities": {
      "get": {
        "summary": "Listar atividades paginadas, com filtros e ordenação",
        "tags": ["Atividades"],
        "description": "Endpoint para listar de forma paginada as atividades registradas. Parâmetros opcionais: type (filtra por um tipo específico de atividades), orderBy (campo pelo qual ordenar as atividades) e order ('asc' = crescente, 'desc' = decrescente).",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Número da página.",
            "required": false,
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Tamanho da página.",
            "required": false,
            "schema": { "type": "integer", "default": 10 }
          },
          {
            "name": "typeId",
            "in": "query",
            "description": "Filtra atividades por tipo.",
            "required": false,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Campo para ordenar as atividades.",
            "required": false,
            "schema": { "type": "string", "default": "createdAt" }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Ordem pela qual ordenar as atividades (asc ou desc).",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "page": { "type": "integer" },
                    "pageSize": { "type": "integer" },
                    "totalActivities": { "type": "integer" },
                    "totalPages": { "type": "integer" },
                    "previous": { "type": "integer" },
                    "next": { "type": "integer" },
                    "activities": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "title": { "type": "string" },
                          "description": { "type": "string" },
                          "type": { "type": "string" },
                          "image": { "type": "string" },
                          "confirmationCode": { "type": "string" },
                          "participantCount": { "type": "integer" },
                          "address": {
                            "type": "object",
                            "properties": {
                              "latitude": { "type": "number" },
                              "longitude": { "type": "number" }
                            }
                          },
                          "scheduledDate": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "completedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "private": { "type": "boolean" },
                          "creator": {
                            "type": "object",
                            "properties": {
                              "id": { "type": "string", "format": "uuid" },
                              "name": { "type": "string" },
                              "avatar": { "type": "string" }
                            }
                          },
                          "userSubscriptionStatus": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/activities/all": {
      "get": {
        "summary": "Listar todas as atividades, com filtro por tipo e ordenação",
        "tags": ["Atividades"],
        "description": "Endpoint para listar de forma paginada as atividades registradas. Parâmetros opcionais: type (filtra por um tipo específico de atividades), orderBy (campo pelo qual ordenar as atividades) e order ('asc' = crescente, 'desc' = decrescente).",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Tipo de atividade.",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Campo para ordenar.",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Ordem (asc ou desc).",
            "required": false,
            "schema": { "type": "string", "enum": ["asc", "desc"] }
          }
        ],
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "title": { "type": "string" },
                      "description": { "type": "string" },
                      "type": { "type": "string" },
                      "image": { "type": "string" },
                      "confirmationCode": { "type": "string" },
                      "participantCount": { "type": "integer" },
                      "address": {
                        "type": "object",
                        "properties": {
                          "latitude": { "type": "number" },
                          "longitude": { "type": "number" }
                        }
                      },
                      "scheduledDate": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "createdAt": { "type": "string", "format": "date-time" },
                      "completedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "private": { "type": "boolean" },
                      "creator": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "name": { "type": "string" },
                          "avatar": { "type": "string" }
                        }
                      },
                      "userSubscriptionStatus": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/activities/user/creator": {
      "get": {
        "summary": "Listar atividades criadas pelo usuário (paginadas)",
        "tags": ["Atividades"],
        "description": "Endpoint para listar de forma paginada as atividades criadas pelo usuário logado.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Número da página.",
            "required": false,
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Tamanho da página.",
            "required": false,
            "schema": { "type": "integer", "default": 10 }
          }
        ],
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "page": { "type": "integer" },
                    "pageSize": { "type": "integer" },
                    "totalActivities": { "type": "integer" },
                    "totalPages": { "type": "integer" },
                    "previous": { "type": "integer" },
                    "next": { "type": "integer" },
                    "activities": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "title": { "type": "string" },
                          "description": { "type": "string" },
                          "type": { "type": "string" },
                          "image": { "type": "string" },
                          "confirmationCode": { "type": "string" },
                          "participantCount": { "type": "integer" },
                          "address": {
                            "type": "object",
                            "properties": {
                              "latitude": { "type": "number" },
                              "longitude": { "type": "number" }
                            }
                          },
                          "scheduledDate": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "completedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "private": { "type": "boolean" },
                          "creator": {
                            "type": "object",
                            "properties": {
                              "id": { "type": "string", "format": "uuid" },
                              "name": { "type": "string" },
                              "avatar": { "type": "string" }
                            }
                          },
                          "userSubscriptionStatus": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/activities/user/creator/all": {
      "get": {
        "summary": "Listar todas as atividades criadas pelo usuário",
        "tags": ["Atividades"],
        "description": "Endpoint para listar todas as atividades criadas pelo usuário logado.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "title": { "type": "string" },
                      "description": { "type": "string" },
                      "type": { "type": "string" },
                      "image": { "type": "string" },
                      "confirmationCode": { "type": "string" },
                      "participantCount": { "type": "integer" },
                      "address": {
                        "type": "object",
                        "properties": {
                          "latitude": { "type": "number" },
                          "longitude": { "type": "number" }
                        }
                      },
                      "scheduledDate": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "createdAt": { "type": "string", "format": "date-time" },
                      "completedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "private": { "type": "boolean" },
                      "creator": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "name": { "type": "string" },
                          "avatar": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/activities/user/participant/all": {
      "get": {
        "summary": "Listar todas as atividades em que o usuário se inscreveu",
        "tags": ["Atividades"],
        "description": "Endpoint para listar todas as atividades em que o usuário logado se inscreveu.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "OK.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "title": { "type": "string" },
                      "description": { "type": "string" },
                      "type": { "type": "string" },
                      "image": { "type": "string" },
                      "confirmationCode": { "type": "string" },
                      "participantCount": { "type": "integer" },
                      "address": {
                        "type": "object",
                        "properties": {
                          "latitude": { "type": "number" },
                          "longitude": { "type": "number" }
                        }
                      },
                      "scheduledDate": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "createdAt": { "type": "string", "format": "date-time" },
                      "completedAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "private": { "type": "boolean" },
                      "creator": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "name": { "type": "string" },
                          "avatar": { "type": "string" }
                        }
                      },
                      "userSubscriptionStatus": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "mensagem": {
                  "type": "string",
                  "example": "Autenticação necessária."
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Esta conta foi desativada e não pode ser utilizada."
                }
              }
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Erro inesperado."
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}